Create abstract stream builder for input/output streams
